calling-assistant/requirements.txt
fastapi 
uvicorn 
twilio 
websockets 
python-dotenv
openai